thermohands-annotator/requirements.txt
absl-py==2.0.0
addict==2.4.0
aiofiles==22.1.0
aiosqlite==0.19.0
annotated-types==0.5.0
ansi2html==1.8.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
attrs==23.1.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
cached-property==1.5.2
cachetools==5.3.2
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi==1.15.1
charset-normalizer==3.3.1
chumpy==0.70
click==8.1.7
cmake==3.27.7
colorama==0.4.6
comm==0.1.4
ConfigArgParse==1.7
cycler==0.11.0
darr==0.5.4
dash==2.14.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
debugpy==1.7.0
decorator==4.4.2
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.7
entrypoints==0.4
exceptiongroup==1.1.3
fastjsonschema==2.18.1
filterpy==1.4.5
Flask==2.2.5
flatbuffers==23.5.26
fonttools==4.38.0
fqdn==1.5.1
fvcore==0.1.5.post20221221
idna==3.4
imagecodecs==2021.11.20
imageio==2.31.2
imageio-ffmpeg==0.4.9
importlib-metadata==6.7.0
importlib-resources==5.12.0
iopath==0.1.10
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.1.1
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server==1.24.0
jupyter-ydoc==0.2.5
jupyter_client==7.4.9
jupyter_core==4.12.0
jupyter_packaging==0.12.3
jupyter_server_fileid==0.9.0
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.24.0
-e git+https://github.com/PRBonn/kiss-icp.git@8e5b7dbafe199068ba2f7e6d52de4867b405e898#egg=kiss_icp&subdirectory=python
kiwisolver==1.4.5
manopth @ file:///home/fangqiang/thermal-hand/manopth
markdown-it-py==2.2.0
MarkupSafe==2.1.3
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.9.0.1
mistune==3.0.2
moviepy==1.0.3
multiprocess==0.70.15
natsort==8.4.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.6.0
nbformat==5.5.0
nest-asyncio==1.5.8
ninja==1.11.1.1
notebook==6.5.6
notebook_shim==0.2.3
numpy==1.21.6
nvidia-ml-py==12.535.133
nvitop==1.3.1
open3d==0.16.0
open3d-cpu==0.17.0
open3d-python==0.7.0.0
opencv-contrib-python==4.8.1.78
opencv-python==4.0.1.23
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
path==16.6.0
pathos==0.3.1
pathspec==0.11.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.17.0
plyfile==0.9
portalocker==2.7.0
pox==0.3.3
ppft==1.7.6.7
proglog==0.1.10
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pybind11==2.11.1
pycparser==2.21
pydantic==2.5.3
pydantic-settings==2.0.3
pydantic_core==2.14.6
Pygments==2.16.1
pyparsing==3.1.1
pyproject-metadata==0.7.1
pyquaternion==0.9.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.7
pytorch3d==0.3.0
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==24.0.1
requests==2.31.0
retrying==1.3.4
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
scikit-learn==1.0.2
scikit_build_core==0.8.0
scipy==1.7.3
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git@6fdee8f2727f4506cfbbe553e23b895e27956588
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
sound==0.1.0
soundfile==0.12.1
soupsieve==2.4.1
spicy==0.16.0
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.1
torch==1.10.0+cu111
torchaudio==0.10.0+rocm4.1
torchvision==0.11.0+cu111
tornado==6.2
tqdm==4.66.1
traitlets==5.9.0
typer==0.9.0
typing_extensions==4.7.1
uri-template==1.3.0
urllib3==2.0.7
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
Werkzeug==2.2.3
widgetsnbextension==4.0.9
y-py==0.6.2
yacs==0.1.8
ypy-websocket==0.8.4
zipp==3.15.0
